fix bug needed to type cast

+++ server/src/main/java/com/tco/misc/CosinesDistance.java
@@ -9,8 +9,8 @@ public class CosinesDistance implements GreatCircleDistance {
         double to_lon = to.lonRadians();
         double to_lat = to.latRadians();
 
-        Long distance = earthRadius * Math.acos(Math.sin(from_lat) * Math.sin(to_lat) + 
-                Math.cos(from_lat) * Math.cos(to_lat) * Math.cos(from_lon - to_lon));
+        Long distance = (long) (earthRadius * Math.acos(Math.sin(from_lat) * Math.sin(to_lat) + 
+                Math.cos(from_lat) * Math.cos(to_lat) * Math.cos(from_lon - to_lon)));
                 
         return distance;
     }
